scripts/requirements.txt
pyalex
ads
pre-commit
python-dotenv